pyproject.toml
[project]
name = "learn-gaussian-splatting-together"
version = "0.1.0"
description = "Learn Gaussian Splatting"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "plyfile>=1.1.2",
    "tqdm>=4.67.1",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "colmap_loader.py",          
    "dataset_readers.py",          
    "spherical.py"
]

line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"
